Implemented repository for auctionProducts

+++ MarketMinds/DataAccessLayer/Repositories/AuctionProductsRepository.cs
@@ -23,78 +23,114 @@ namespace DataAccessLayer.Repositories
         {
             List<Product> auctions = new List<Product>();
 
-        //    string query = @"
-        //SELECT 
-        //    ap.id,
-        //    ap.description,
-        //    ap.seller_id,
-        //    u.username,
-        //    u.email,
-        //    ap.condition_id,
-        //    pc.title AS conditionTitle,
-        //    pc.description AS conditionDescription,
-        //    ap.category_id,
-        //    cat.title AS categoryTitle,
-        //    cat.description AS categoryDescription,
-        //    ap.start_datetime,
-        //    ap.end_datetime,
-        //    ap.starting_price
-        //FROM AuctionProducts ap
-        //JOIN Users u ON ap.seller_id = u.id
-        //JOIN ProductConditions pc ON ap.condition_id = pc.id
-        //JOIN ProductCategories cat ON ap.category_id = cat.id";
+            string query = @"
+        SELECT 
+            ap.id,
+            ap.title,
+            ap.description,
+            ap.seller_id,
+            u.username,
+            u.email,
+            ap.condition_id,
+            pc.title AS conditionTitle,
+            pc.description AS conditionDescription,
+            ap.category_id,
+            cat.title AS categoryTitle,
+            cat.description AS categoryDescription,
+            ap.start_datetime,
+            ap.end_datetime,
+            ap.starting_price
+        FROM AuctionProducts ap
+        JOIN Users u ON ap.seller_id = u.id
+        JOIN ProductConditions pc ON ap.condition_id = pc.id
+        JOIN ProductCategories cat ON ap.category_id = cat.id";
 
-        //    connection.OpenConnection();
-        //    using (SqlCommand cmd = new SqlCommand(query, connection.GetConnection()))
-        //    {
-        //        using (SqlDataReader reader = cmd.ExecuteReader())
-        //        {
-        //            while (reader.Read())
-        //            {
-        //                int id = reader.GetInt32(reader.GetOrdinal("id"));
-        //                string description = reader.GetString(reader.GetOrdinal("description"));
+            connection.OpenConnection();
+            using (SqlCommand cmd = new SqlCommand(query, connection.GetConnection()))
+            {
+                using (SqlDataReader reader = cmd.ExecuteReader())
+                {
+                    while (reader.Read())
+                    {
+                        int id = reader.GetInt32(reader.GetOrdinal("id"));
+                        string title = reader.GetString(reader.GetOrdinal("title"));
+                        string description = reader.GetString(reader.GetOrdinal("description"));
 
-        //                int sellerId = reader.GetInt32(reader.GetOrdinal("seller_id"));
-        //                string username = reader.GetString(reader.GetOrdinal("username"));
-        //                string email = reader.GetString(reader.GetOrdinal("email"));
-        //                User seller = new User(sellerId, username, email);
+                        int sellerId = reader.GetInt32(reader.GetOrdinal("seller_id"));
+                        string username = reader.GetString(reader.GetOrdinal("username"));
+                        string email = reader.GetString(reader.GetOrdinal("email"));
+                        User seller = new User(sellerId, username, email);
 
-                      
-        //                int conditionId = reader.GetInt32(reader.GetOrdinal("condition_id"));
-        //                string conditionTitle = reader.GetString(reader.GetOrdinal("conditionTitle"));
-        //                string conditionDescription =  reader.GetString(reader.GetOrdinal("conditionDescription"));
-        //                ProductCondition condition = new ProductCondition(conditionId, conditionTitle, conditionDescription);
 
-                       
-        //                int categoryId = reader.GetInt32(reader.GetOrdinal("category_id"));
-        //                string categoryTitle = reader.GetString(reader.GetOrdinal("categoryTitle"));
-        //                string categoryDescription =  reader.GetString(reader.GetOrdinal("categoryDescription"));
-        //                ProductCategory category = new ProductCategory(categoryId, categoryTitle, categoryDescription);
+                        int conditionId = reader.GetInt32(reader.GetOrdinal("condition_id"));
+                        string conditionTitle = reader.GetString(reader.GetOrdinal("conditionTitle"));
+                        string conditionDescription = reader.GetString(reader.GetOrdinal("conditionDescription"));
+                        ProductCondition condition = new ProductCondition(conditionId, conditionTitle, conditionDescription);
 
-        //                DateTime start = reader.GetDateTime(reader.GetOrdinal("start_datetime"));
-        //                DateTime end = reader.GetDateTime(reader.GetOrdinal("end_datetime"));
-        //                float startingPrice = reader.GetFloat(reader.GetOrdinal("starting_price"));
 
-        //                List<ProductTag> tags = GetProductTags(id); 
+                        int categoryId = reader.GetInt32(reader.GetOrdinal("category_id"));
+                        string categoryTitle = reader.GetString(reader.GetOrdinal("categoryTitle"));
+                        string categoryDescription = reader.GetString(reader.GetOrdinal("categoryDescription"));
+                        ProductCategory category = new ProductCategory(categoryId, categoryTitle, categoryDescription);
 
-        //                AuctionProduct auction = new AuctionProduct(
-        //                    id,
-        //                    description,
-        //                    seller,
-        //                    condition,
-        //                    category,
-        //                    tags,
-        //                    start,
-        //                    end,
-        //                    startingPrice
-        //                );
+                        DateTime start = reader.GetDateTime(reader.GetOrdinal("start_datetime"));
+                        DateTime end = reader.GetDateTime(reader.GetOrdinal("end_datetime"));
+                        float startingPrice = reader.GetFloat(reader.GetOrdinal("starting_price"));
 
-        //                auctions.Add(auction);
-        //            }
-        //        }
-        //    }
-        //    connection.CloseConnection();
+                        List<ProductTag> tags = GetProductTags(id);
+
+                        List<Image> images = GetImages(id);
+
+                        AuctionProduct auction = new AuctionProduct(
+                            id,
+                            title,
+                            description,
+                            seller,
+                            condition,
+                            category,
+                            tags,
+                            images,
+                            start,
+                            end,
+                            startingPrice
+                        );
+
+                        auctions.Add(auction);
+                    }
+                }
+            }
+            connection.CloseConnection();
             return auctions;
+        }
+
+        private List<Image> GetImages(int productId)
+        {
+            var images = new List<Image>();
+
+            string query = @"
+            SELECT url
+            FROM AuctionProductsImages
+            WHERE product_id = @ProductId";
+
+            connection.OpenConnection();
+            using (SqlCommand cmd = new SqlCommand(query, connection.GetConnection()))
+            {
+                cmd.Parameters.AddWithValue("@ProductId", productId);
+
+                using (SqlDataReader reader = cmd.ExecuteReader())
+                {
+                    while (reader.Read())
+                    {
+                        int imageId = reader.GetInt32(reader.GetOrdinal("id"));
+                        string url = reader.GetString(reader.GetOrdinal("url"));
+
+                        images.Add(new Image(url));
+                    }
+                }
+            }
+            connection.CloseConnection();
+
+            return images;
         }
 
         private List<ProductTag> GetProductTags(int productId)
@@ -130,13 +166,17 @@ namespace DataAccessLayer.Repositories
 
         public override void DeleteProduct(Product product)
         {
+            AuctionProduct auction = product as AuctionProduct;
+            if (auction == null)
+                throw new ArgumentException("Product must be of type AuctionProduct.");
+
             string query = "DELETE FROM AuctionProducts WHERE id = @Id";
 
             connection.OpenConnection();
             using (SqlCommand cmd = new SqlCommand(query, connection.GetConnection()))
 
             {
-                cmd.Parameters.AddWithValue("@Id", product.Id);
+                cmd.Parameters.AddWithValue("@Id", auction.Id);
 
                 cmd.ExecuteNonQuery();
             }
@@ -144,29 +184,171 @@ namespace DataAccessLayer.Repositories
 
         public override void AddProduct(Product product)
         {
-            throw new NotImplementedException();
+            AuctionProduct auction = (AuctionProduct)product;
+            if (auction == null)
+                throw new ArgumentException("Product must be of type AuctionProduct.");
 
+
+            string insertProductQuery = @"
+            INSERT INTO AuctionProducts 
+            (description, seller_id, condition_id, category_id, start_datetime, end_datetime, starting_price , title)
+            VALUES 
+            (@Description, @SellerId, @ConditionId, @CategoryId, @StartDateTime, @EndDateTime, @StartingPrice , @Title);
+            SELECT SCOPE_IDENTITY();";
+
+            connection.OpenConnection();
+
+            int newProductId;
+            using (SqlCommand cmd = new SqlCommand(insertProductQuery, connection.GetConnection()))
+            {                
+                cmd.Parameters.AddWithValue("@Description", auction.Description);
+                cmd.Parameters.AddWithValue("@SellerId", auction.Seller.Id);
+                cmd.Parameters.AddWithValue("@ConditionId", auction.Condition.id);
+                cmd.Parameters.AddWithValue("@CategoryId", auction.Category.id);
+                cmd.Parameters.AddWithValue("@StartDateTime", auction.StartAuctionDate);
+                cmd.Parameters.AddWithValue("@EndDateTime", auction.EndAuctionDate);
+                cmd.Parameters.AddWithValue("@StartingPrice", auction.StartingPrice);
+                cmd.Parameters.AddWithValue("@Title", auction.Title);
+
+                object result = cmd.ExecuteScalar();
+                newProductId = Convert.ToInt32(result);
+            }
+
+            foreach (var tag in auction.Tags)
+            {
+                string insertTagQuery = @"
+            INSERT INTO AuctionProductProductTags (product_id, tag_id)
+            VALUES (@ProductId, @TagId)";
+
+                using (SqlCommand cmd = new SqlCommand(insertTagQuery, connection.GetConnection()))
+                {
+                    cmd.Parameters.AddWithValue("@ProductId", newProductId);
+                    cmd.Parameters.AddWithValue("@TagId", tag.id);
+                    cmd.ExecuteNonQuery();
+                }
+            }
+
+            foreach (var image in auction.Images)
+            {
+                string insertImageQuery = @"
+            INSERT INTO AuctionProductsImages (url, product_id)
+            VALUES (@Url, @ProductId)";
+
+                using (SqlCommand cmd = new SqlCommand(insertImageQuery, connection.GetConnection()))
+                {
+                    cmd.Parameters.AddWithValue("@Url", image.url);
+                    cmd.Parameters.AddWithValue("@ProductId", newProductId);
+                    cmd.ExecuteNonQuery();
+                }
+            }
+
+            connection.CloseConnection();
         }
 
 
         public override void UpdateProduct(Product product)
         {
-            //string query = "UPDATE AuctionProducts SET current_price = @CurrentPrice WHERE Id = @Id";
-            //connection.OpenConnection();
-            //using (SqlCommand cmd = new SqlCommand(query, connection.GetConnection()))
-            //{
-            //    cmd.Parameters.AddWithValue("@CurrentPrice", product.CurrentPrice);
-            //    cmd.Parameters.AddWithValue("@Id", product.Id);
+            AuctionProduct auction = (AuctionProduct)product;
+            if (auction == null)
+                throw new ArgumentException("Product must be of type AuctionProduct.");
 
-                
-            //    cmd.ExecuteNonQuery();
-            //}
+            string query = "UPDATE AuctionProducts SET current_price = @CurrentPrice WHERE Id = @Id";
+
+            connection.OpenConnection();
+            using (SqlCommand cmd = new SqlCommand(query, connection.GetConnection()))
+            {
+                cmd.Parameters.AddWithValue("@CurrentPrice", auction.CurrentPrice);
+                cmd.Parameters.AddWithValue("@Id", auction.Id);
+
+
+                cmd.ExecuteNonQuery();
+            }
         }
 
-        public override AuctionProduct GetProductByID(int id)
+        public override AuctionProduct GetProductByID(int APid)
         {
-            throw new NotImplementedException();
-          
+            AuctionProduct auction = null;
+
+            string query = @"
+            SELECT 
+                ap.id,
+                ap.title,
+                ap.description,
+                ap.seller_id,
+                u.username,
+                u.email,
+                ap.condition_id,
+                pc.title AS conditionTitle,
+                pc.description AS conditionDescription,
+                ap.category_id,
+                cat.title AS categoryTitle,
+                cat.description AS categoryDescription,
+                ap.start_datetime,
+                ap.end_datetime,
+                ap.starting_price
+            FROM AuctionProducts ap
+            JOIN Users u ON ap.seller_id = u.id
+            JOIN ProductConditions pc ON ap.condition_id = pc.id
+            JOIN ProductCategories cat ON ap.category_id = cat.id
+               WHERE ap.id = @APid";
+
+            connection.OpenConnection();
+            using (SqlCommand cmd = new SqlCommand(query, connection.GetConnection()))
+            {
+                cmd.Parameters.AddWithValue("@APid", APid);
+                using (SqlDataReader reader = cmd.ExecuteReader())
+                {
+                    while (reader.Read())
+                    {
+                        int id = reader.GetInt32(reader.GetOrdinal("id"));
+                        string title = reader.GetString(reader.GetOrdinal("title"));
+                        string description = reader.GetString(reader.GetOrdinal("description"));
+
+                        int sellerId = reader.GetInt32(reader.GetOrdinal("seller_id"));
+                        string username = reader.GetString(reader.GetOrdinal("username"));
+                        string email = reader.GetString(reader.GetOrdinal("email"));
+                        User seller = new User(sellerId, username, email);
+
+
+                        int conditionId = reader.GetInt32(reader.GetOrdinal("condition_id"));
+                        string conditionTitle = reader.GetString(reader.GetOrdinal("conditionTitle"));
+                        string conditionDescription = reader.GetString(reader.GetOrdinal("conditionDescription"));
+                        ProductCondition condition = new ProductCondition(conditionId, conditionTitle, conditionDescription);
+
+
+                        int categoryId = reader.GetInt32(reader.GetOrdinal("category_id"));
+                        string categoryTitle = reader.GetString(reader.GetOrdinal("categoryTitle"));
+                        string categoryDescription = reader.GetString(reader.GetOrdinal("categoryDescription"));
+                        ProductCategory category = new ProductCategory(categoryId, categoryTitle, categoryDescription);
+
+                        DateTime start = reader.GetDateTime(reader.GetOrdinal("start_datetime"));
+                        DateTime end = reader.GetDateTime(reader.GetOrdinal("end_datetime"));
+                        float startingPrice = reader.GetFloat(reader.GetOrdinal("starting_price"));
+
+                        List<ProductTag> tags = GetProductTags(id);
+
+                        List<Image> images = GetImages(id);
+
+                        auction = new AuctionProduct(
+                            id,
+                            title,
+                            description,
+                            seller,
+                            condition,
+                            category,
+                            tags,
+                            images,
+                            start,
+                            end,
+                            startingPrice
+                        );
+                    }
+                }
+            }
+            connection.CloseConnection();
+
+            return auction;
         }
+
     }
 }
